requirements.txt
asyncio==3.4.3
minimalmodbus==2.1.1
pymodbus==3.6.9
pyserial==3.5
